* change index/buffer stuff from int to short

+++ branches/new/mt4j-desktop/extensions/org/mt4jx/util/extension3D/collision/CollisionManager.java
@@ -201,7 +201,7 @@ public class CollisionManager {
 	 */
 	public ByteBuffer getIndexBuffer(MTTriangleMesh mesh)
 	{
-		int[] indices = mesh.getGeometryInfo().getIndices();
+		short[] indices = mesh.getGeometryInfo().getIndices();
 		ByteBuffer buf = ByteBuffer.allocateDirect(indices.length*4).order(ByteOrder.nativeOrder());
 		
 		for (int i=0; i<indices.length; i++) {
